Move spec helper into Test module which is not required by default
