+++ lib/phobos.rb
@@ -22,7 +22,6 @@ require 'phobos/producer'
 require 'phobos/handler'
 require 'phobos/echo_handler'
 require 'phobos/executor'
-require 'phobos/spec_helper'
 
 Thread.abort_on_exception = true
