+++ lib/phobos/test.rb
@@ -1,0 +1,1 @@
+require 'phobos/test/helper'
