+++ spec/spec_helper.rb
@@ -21,6 +21,7 @@ SimpleCov.start
 require 'phobos'
 require 'pry-byebug'
 require 'timecop'
+require 'phobos/test'
 
 Dir.entries('./spec/support').select { |f| f =~ /\.rb$/ }.each do |f|
   load "./spec/support/#{f}"
